+++ addOns/reports/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Added
 - Wappalyzer data to the traditional-html-plus report, if it is available.
+- Traditional JSON report
 
 ### Changed
 - Maintenance changes.